+++ src/index.js
@@ -44,29 +44,29 @@
 import "./js/helpers/darkThema";
 import "./js/helpers/cursor";
 import apiService from "./js/apiService";
-import renderGallery from "./js/components/startPage/movieGallary";
-import { renderPaginationButtons } from "./js/components/pagination/pagination";
-import renderMovieModal from "./js/components/modal/renderMovieModal";
-import openModal from "./js/components/modal/modal-open";
-import onSearchButton from "./js/components/search/movieSearch";
+import renderGallery from "./js/components/movieGallary";
+import { renderPaginationButtons } from "./js/pagination";
+import renderMovieModal from "./js/components/renderMovieModal";
+import openModal from "./js/modal-open";
+import onSearchButton from "./js/movieSearch";
 
 import NProgress from 'nprogress';
 
 import './js/team-modal';
-import './js/components/authorization/auth-page-close'
+import './js/auth-page-close'
 import './js/helpers/sliderTrendingMovies';
 import { renderMovieCardsSlider } from './js/templates/renderMovieCardsSlider';
 
 import './js/helpers/back-to-top';
 
-import * as auth from './js/components/authorization/auth';
+import * as auth from './js/auth';
 
 
 import {markupGallery} from './js/templates/startPageMarkup'
-import { watchedArr, queueArr } from './js/components/modal/modal-open'
+import { watchedArr, queueArr } from './js/modal-open'
 
-import addToLibrary from './js/components/library/addToLibrary';
-import removeFromLibrary from './js/components/library/removeFromLibrary';
+import addToLibrary from './js/addToLibrary';
+import removeFromLibrary from './js/removeFromLibrary';
