+++ src/js/addToLibrary.js
