+++ src/js/auth-page-close.js
