+++ src/js/components/movieGallary.js
@@ -1,5 +1,5 @@
 
-import noPoster from '../../../images/no-poster.jpg';
+import noPoster from '../../images/no-poster.jpg';
 const IMG_URL= `https://image.tmdb.org/t/p/w500`;
