+++ src/js/library.js
@@ -1,10 +1,10 @@
-import '../../helpers/darkThema';
-import '../../helpers/cursor';
-import openModal from '../modal/modal-open';
-import renderLibrary from '../../templates/movieCardLibrary';
-import { renderPaginationBTN, resetPagination } from '../pagination/paginationLibrary';
-import '../../team-modal';
-import app from '../../helpers/initialize-firebase';
+import './helpers/darkThema';
+import './helpers/cursor';
+import openModal from './modal-open';
+import renderLibrary from './templates/movieCardLibrary';
+import { renderPaginationBTN, resetPagination } from './paginationLibrary';
+import './team-modal';
+import app from './helpers/initialize-firebase';
 import { signOut, getAuth } from 'firebase/auth';
 
 let watched = [];
