+++ src/js/removeFromLibrary.js
