+++ src/library.html
@@ -11,7 +11,7 @@
       rel="stylesheet"
     />
     <link rel="stylesheet" href="./sass/index.scss" />
-    <script type="module" src="./js/components/library/library.js"></script>
+    <script type="module" src="./js/library.js"></script>
   </head>
 
   <body class="library">
